change lenght from integer to biginteger

+++ database/migrations/2023_12_21_152025_create_temuans_tables.php
@@ -18,10 +18,10 @@ return new class extends Migration
             $table->integer('lhp_id');
             $table->string('jns_pemeriksaan');
             $table->string('ringkasan');
-            $table->integer('nilai_temuan');
+            $table->bigInteger('nilai_temuan');
             $table->string('jns_temuan');
             $table->string('rekomendasi');
-            $table->integer('nilai_rekomendasi');
+            $table->bigInteger('nilai_rekomendasi');
             $table->timestamps();
             $table->softDeletes();
         });
