+++ database/migrations/2024_01_22_143036_add_nilai_temuan_to_tindak_lanjuts.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('tindak_lanjuts', function (Blueprint $table) {
-            $table->integer('nilai_temuan')->after('lhp_id')->nullable();
+            $table->bigInteger('nilai_temuan')->after('lhp_id')->nullable();
         });
     }
